+++ lib/SILGen/CMakeLists.txt
@@ -4,6 +4,7 @@ add_swift_library(swiftSILGen STATIC
   Condition.cpp
   FormalEvaluation.cpp
   ManagedValue.cpp
+  ResultPlan.cpp
   RValue.cpp
   SwitchCaseFullExpr.cpp
   SILGen.cpp
